+++ src/pages/GraduationCalculatorPage/components/CourseTable/CourseTable.module.scss
@@ -11,7 +11,7 @@
   gap: 26px;
 }
 
-.table {
+.content {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
@@ -20,106 +20,33 @@
   font-size: 14px;
   font-weight: 400;
   line-height: 22.4px;
+}
 
-  &__content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 264px;
-    gap: 4px;
+.trigger {
+  display: flex;
+  width: 100%;
+  height: 60px;
+  padding: 0 12px;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+  position: relative;
+  bottom: 0;
+  box-sizing: border-box;
+  background-color: #fff;
+  border: 1px solid #e1e1e1;
+  border-radius: 16px;
+  font-family: Pretendard, sans-serif;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 160%;
+
+  &:hover {
+    background-color: #eee;
   }
 
-  &__header {
-    display: flex;
-    width: 100%;
-    box-sizing: border-box;
-    padding: 0 12px;
-    align-items: center;
-    background-color: #f1f8ff;
-    color: #4b4b4b;
-
-    & > tr > th:nth-child(1) { width: 138px; }
-
-    & > tr > th:nth-child(2) { width: 48px; }
-
-    & > tr > th:nth-child(3) { width: 37px; }
-
-    & > tr > th:nth-child(4) { width: 106px; }
-
-    & > tr > th:nth-child(5) { width: 16px; }
-  }
-
-  &__body {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    box-sizing: border-box;
-    padding: 0 12px;
-    gap: 4px;
-    align-items: center;
-    align-self: stretch;
-    color: #000;
-    overflow-y: auto;
-
-    & > tr > td:nth-child(1) { width: 138px; }
-
-    & > tr > td:nth-child(2) { width: 48px; }
-
-    & > tr > td:nth-child(3) { width: 37px; }
-
-    & > tr > td:nth-child(4) { width: 106px; }
-
-    & > tr > td:nth-child(5) {
-      width: 16px;
-
-      button {
-        border-radius: 4px;
-        width: 16px;
-        height: 16px;
-
-        &:hover {
-          background-color: #eee;
-        }
-
-        &:active {
-          background-color: #e1e1e1;
-        }
-      }
-
-      svg {
-        width: 8px;
-        height: 8px;
-        vertical-align: middle;
-      }
-    }
-  }
-
-  &__trigger {
-    display: flex;
-    width: 100%;
-    height: 60px;
-    padding: 0 12px;
-    justify-content: center;
-    align-items: center;
-    gap: 20px;
-    position: relative;
-    bottom: 0;
-    box-sizing: border-box;
-    background-color: #fff;
-    border: 1px solid #e1e1e1;
-    border-radius: 16px;
-    font-family: Pretendard, sans-serif;
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 160%;
-
-    &:hover {
-      background-color: #eee;
-    }
-
-    &:active {
-      background-color: #e1e1e1;
-    }
+  &:active {
+    background-color: #e1e1e1;
   }
 }
