+++ src/pages/GraduationCalculatorPage/components/CourseTable/SemesterCourseTable/SemesterCourseTable.module.scss
@@ -1,0 +1,79 @@
+@use "utils/scss/media.scss" as media;
+
+
+.table {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: 264px;
+  gap: 4px;
+
+  ::-webkit-scrollbar {
+    display: none;
+  }
+
+  &__header {
+    display: flex;
+    width: 100%;
+    box-sizing: border-box;
+    padding: 0 12px;
+    align-items: center;
+    background-color: #f1f8ff;
+    color: #4b4b4b;
+  
+    & > tr > th:nth-child(1) { width: 138px; }
+  
+    & > tr > th:nth-child(2) { width: 48px; }
+  
+    & > tr > th:nth-child(3) { width: 37px; }
+  
+    & > tr > th:nth-child(4) { width: 106px; }
+  
+    & > tr > th:nth-child(5) { width: 16px; }
+  }
+  
+  &__body {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    box-sizing: border-box;
+    padding: 0 12px;
+    gap: 4px;
+    align-items: center;
+    align-self: stretch;
+    color: #000;
+    overflow-y: auto;
+  
+    & > tr > td:nth-child(1) { width: 138px; }
+  
+    & > tr > td:nth-child(2) { width: 48px; }
+  
+    & > tr > td:nth-child(3) { width: 37px; }
+  
+    & > tr > td:nth-child(4) { width: 106px; }
+  
+    & > tr > td:nth-child(5) {
+      width: 16px;
+  
+      button {
+        border-radius: 4px;
+        width: 16px;
+        height: 16px;
+  
+        &:hover {
+          background-color: #eee;
+        }
+  
+        &:active {
+          background-color: #e1e1e1;
+        }
+      }
+  
+      svg {
+        width: 8px;
+        height: 8px;
+        vertical-align: middle;
+      }
+    }
+  }
+}
